MODLD-81: fix permission, refactor ISBN processor and fix Standalone mode
